Make register op implementation constant
