+++ tfjs-tfdf/src/setup_test.ts
@@ -17,5 +17,7 @@
 import '@tensorflow/tfjs-backend-cpu';
 
 import './setup_test_globals';
+// Register ops.
+import './index';
 import './tfdf_web_api_client_test';
 import './tfdf_model_test';
